refactor(PrintMeesage, ErrorMessage) : 메세지들을 messages로 이동시킨다
기능을 명시할 수 있도록 메세지 클래스들을 messages 폴더로 이동한다
